Rename: rts.* -> ts.*
