+++ core/fts.h
@@ -19,7 +19,7 @@
 
 #include <unordered_map>
 
-#include "rts.h"
+#include "ts.h"
 
 namespace cosa {
