+++ core/ts.h
